solving nans

+++ content_funcs.h
@@ -19,18 +19,18 @@ using namespace std;
 typedef struct item_type{
 	int year;
 	int runtime;
-	float imdb_rating;
+	double imdb_rating;
 	unordered_map<string, int> word_freq;
 	string director;
 	string language;
 
 	/*--- This variable was created in order to save time when calculating the similarities ---*/
-	float norm;
+	double norm;
 
 	item_type(){}
 
-	item_type(int _year, int _runtime, float _imdb_rating, unordered_map<string, int> _word_freq, 
-				string _director, string _language, float _norm){
+	item_type(int _year, int _runtime, double _imdb_rating, unordered_map<string, int> _word_freq, 
+				string _director, string _language, double _norm){
 		year = _year;
 		runtime = _runtime;
 		imdb_rating = _imdb_rating;
@@ -58,13 +58,13 @@ vector<string> tokenize(string plot);
 unordered_map<string, int> count_freq(set<string> stopwords, vector<string> tokens);
 
 /*--- This function receives a vector in a dense representation and calculate its norm ---*/
-float calc_norm(unordered_map<string, int> v);
+double calc_norm(unordered_map<string, int> v);
 
-float cos_similarity(item_type item1, item_type item2);
+double cos_similarity(item_type item1, item_type item2);
 
-float absolute_value(float x);
+double absolute_value(double x);
 
 /*--- Returns a similarity measure that lays in [0-1] and considers the common features and the textual features ---*/
-float sim(item_type item1, item_type item2);
+double sim(item_type item1, item_type item2);
 
 unordered_map<string, item_type> read_content(set<string> stopwords, const char *contents_file);
