+++ ratings.h
@@ -10,7 +10,7 @@ using namespace std;
 
 unordered_map<string, unordered_map<string, int> > read_ratings(const char *ratings_file);
 
-float predict(unordered_map<string, item_type> &items, unordered_map<string, int> &consumed_items, string item_unknown);
+double predict(unordered_map<string, item_type> &items, unordered_map<string, int> &consumed_items, string item_unknown);
 
 void build_submission_file(unordered_map<string, item_type> items, 
 							unordered_map<string, unordered_map<string, int> > user_to_items,
